[Silver III] Title: 주유소, Time: 340 ms, Memory: 38768 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -3,8 +3,6 @@
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.util.SimpleTimeZone;
-import java.util.StringTokenizer;
 
 public class Main {
     public static void main(String[] args) throws IOException {
@@ -12,22 +10,24 @@ public class Main {
         // 도시 개수
         int N = Integer.parseInt(br.readLine());
 
+        // 도시 간 거리 정보 입력
         String[] stringLenArr = br.readLine().split(" ");
         int[] lenArr = new int[N-1];
         for(int i=0;i<N-1;i++){
             lenArr[i] = Integer.parseInt(stringLenArr[i]);
         }
 
+        // 도시 별 가스 가격 정보 입력
         String[] stringCostArr = br.readLine().split(" ");
         int[] costArr = new int[N];
         for(int i=0;i<N;i++) {
             costArr[i] = Integer.parseInt(stringCostArr[i]);
         }
 
-        int result = 0;
+        long result = 0;
 
-        int currentCost = costArr[0];
-        int tempLenSum = lenArr[0];
+        long currentCost = costArr[0];
+        long tempLenSum = lenArr[0];
 
         for(int i=1;i<N-1;i++){
             if(costArr[i]>=currentCost) {
